chore(cli): split processes into system and environment-info
