+++ packages/botonic-cli/src/util/environment-info.ts
@@ -1,44 +1,10 @@
-import { execSync, spawn } from 'child_process'
 import { arch, platform } from 'os'
 import { join } from 'path'
 
 import { BOTONIC_NPM_NAMESPACE } from '../constants'
 import { SystemInformation } from '../interfaces'
 import { readJSON } from './file-system'
-
-export async function sleep(ms: number): Promise<number> {
-  return new Promise(resolve => setTimeout(resolve, ms))
-}
-
-export function execCommand(command: string): string {
-  try {
-    return String(execSync(command)).trim()
-  } catch (e) {
-    return String(e)
-  }
-}
-
-export function spawnProcess(
-  command: string,
-  args: string[],
-  onClose?: () => string
-): void {
-  const childProcess = spawn(command, args, { shell: true }) // https://nodejs.org/api/child_process.html#child_process_spawning_bat_and_cmd_files_on_windows
-  childProcess.stdout.on('data', out => {
-    process.stdout.write(out)
-  })
-  childProcess.stderr.on('data', stderr => {
-    process.stderr.write(stderr)
-  })
-  childProcess.on('close', code => {
-    onClose && onClose()
-    process.stdout.write(`child process exited with code ${code}`)
-  })
-}
-
-export function spawnNpmScript(script: string, onClose?: () => string): void {
-  spawnProcess('npm', ['run', script], onClose)
-}
+import { execCommand } from './system'
 
 export function isWindows(): boolean {
   return platform() === 'win32'
